Add cirq.eject_phased_paulis transformer to replace cirq.EjectPhasedPaulis
