+++ cirq-core/cirq/__init__.py
@@ -362,6 +362,7 @@ from cirq.transformers import (
     decompose_two_qubit_interaction_into_four_fsim_gates,
     drop_empty_moments,
     drop_negligible_operations,
+    eject_phased_paulis,
     expand_composite,
     is_negligible_turn,
     map_moments,
