+++ cirq-core/cirq/optimizers/merge_interactions_to_sqrt_iswap_test.py
@@ -39,7 +39,6 @@ def assert_optimizes(before: cirq.Circuit, expected: cirq.Circuit, **kwargs):
     # Ignore differences that would be caught by follow-up optimizations.
     followup_optimizations: List[Callable[[cirq.Circuit], None]] = [
         cirq.merge_single_qubit_gates_into_phased_x_z,
-        cirq.EjectPhasedPaulis().optimize_circuit,
         cirq.EjectZ().optimize_circuit,
     ]
     for post in followup_optimizations:
@@ -47,6 +46,7 @@ def assert_optimizes(before: cirq.Circuit, expected: cirq.Circuit, **kwargs):
         post(expected)
 
     followup_transformers: List[cirq.TRANSFORMER] = [
+        cirq.eject_phased_paulis,
         cirq.drop_negligible_operations,
         cirq.drop_empty_moments,
     ]
