+++ cirq-core/cirq/transformers/analytical_decompositions/two_qubit_to_cz.py
@@ -23,9 +23,9 @@ from cirq.linalg.decompositions import num_cnots_required, extract_right_diag
 
 from cirq import ops, linalg, protocols, circuits
 from cirq.transformers.analytical_decompositions import single_qubit_decompositions
+from cirq.transformers.eject_phased_paulis import eject_phased_paulis
 from cirq.optimizers import (
     eject_z,
-    eject_phased_paulis,
     merge_single_qubit_gates,
 )
 
@@ -164,7 +164,7 @@ def _xx_yy_zz_interaction_via_full_czs(
 def _cleanup_operations(operations: Sequence[ops.Operation]):
     circuit = circuits.Circuit(operations)
     merge_single_qubit_gates.merge_single_qubit_gates_into_phased_x_z(circuit)
-    eject_phased_paulis.EjectPhasedPaulis().optimize_circuit(circuit)
+    circuit = eject_phased_paulis(circuit)
     eject_z.EjectZ().optimize_circuit(circuit)
     circuit = circuits.Circuit(circuit.all_operations(), strategy=circuits.InsertStrategy.EARLIEST)
     return list(circuit.all_operations())
